pyproject.toml
[project]
name = "rag-pipeline-classic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "langchain-openai>=0.3.35",
    "pinecone>=7.3.0",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
]
